add update quantity function
