+++ index.html
@@ -59,7 +59,7 @@
               </li>
               <li class="header-icons-item">
                 <a class="link-cart" href="">
-                  <!-- <span class="cart-quantity"></span> -->
+                  <span id="cart-quantity"></span>
                   <i class="icon icon-cart"></i>
                 </a>
               </li>
