+++ js/index.js
@@ -1,3 +1,4 @@
+import { updateCartQty } from "./addCart.js";
 import {
   arrivedProductsArr as arrivedProduct,
   productsDataArr as productsData,
@@ -6,6 +7,6 @@ import {
 import { createProductList, displayProducts } from "./renderProducts.js";
 //get container
 
-
+updateCartQty()
 displayProducts(productsData, "category-product");
 displayProducts(arrivedProduct, "new-arrived-product");
