+++ stylesheet/scss/layout/_nav.scss
@@ -17,7 +17,7 @@
     display: relative;
   }
 }
-.cart-quantity {
+#cart-quantity {
   position: absolute;
   display: inline-block;
   background-color: red;
@@ -28,6 +28,7 @@
   top: 0;
   left: 0;
   transform: translate(-50%,-70%);
+  display: none;
 }
 
 @media only screen and (max-width: 576px) {
